fix unit tests, simplify
