--- packages/rich-text/src/component/test/index.js
@@ -1,116 +1,0 @@
-/**
- * Internal dependencies
- */
-import { diffAriaProps, pickAriaProps } from '../aria';
-
-describe( 'RichText', () => {
-	describe( 'pickAriaProps()', () => {
-		it( 'should should filter all properties to only those begining with "aria-"', () => {
-			expect( pickAriaProps( {
-				tagName: 'p',
-				className: 'class1 class2',
-				'aria-label': 'my label',
-				style: {
-					backgroundColor: 'white',
-					color: 'black',
-					fontSize: '12px',
-					textAlign: 'left',
-				},
-				'aria-owns': 'some-id',
-				'not-aria-prop': 'value',
-				ariaWithoutDash: 'value',
-			} ) ).toEqual( {
-				'aria-label': 'my label',
-				'aria-owns': 'some-id',
-			} );
-		} );
-	} );
-	describe( 'diffAriaProps()', () => {
-		it( 'should report empty arrays for no props', () => {
-			expect( diffAriaProps( {}, {} ) ).toEqual( {
-				removedKeys: [],
-				updatedKeys: [],
-			} );
-		} );
-		it( 'should report empty arrays for non-aria props', () => {
-			expect( diffAriaProps( {
-				'non-aria-prop': 'old value',
-				'removed-prop': 'value',
-			}, {
-				'non-aria-prop': 'new value',
-				'added-prop': 'value',
-			} ) ).toEqual( {
-				removedKeys: [],
-				updatedKeys: [],
-			} );
-		} );
-		it( 'should report added aria props', () => {
-			expect( diffAriaProps( {
-			}, {
-				'aria-prop': 'value',
-			} ) ).toEqual( {
-				removedKeys: [],
-				updatedKeys: [ 'aria-prop' ],
-			} );
-		} );
-		it( 'should report removed aria props', () => {
-			expect( diffAriaProps( {
-				'aria-prop': 'value',
-			}, {
-			} ) ).toEqual( {
-				removedKeys: [ 'aria-prop' ],
-				updatedKeys: [],
-			} );
-		} );
-		it( 'should report changed aria props', () => {
-			expect( diffAriaProps( {
-				'aria-prop': 'old value',
-			}, {
-				'aria-prop': 'new value',
-			} ) ).toEqual( {
-				removedKeys: [],
-				updatedKeys: [ 'aria-prop' ],
-			} );
-		} );
-		it( 'should not report unchanged aria props', () => {
-			expect( diffAriaProps( {
-				'aria-prop': 'value',
-			}, {
-				'aria-prop': 'value',
-			} ) ).toEqual( {
-				removedKeys: [],
-				updatedKeys: [],
-			} );
-		} );
-		it( 'should work with a mixture of aria and non-aria props', () => {
-			expect( diffAriaProps( {
-				tagName: 'p',
-				className: 'class1 class2',
-				'aria-label': 'my label',
-				style: {
-					backgroundColor: 'white',
-					color: 'black',
-					fontSize: '12px',
-					textAlign: 'left',
-				},
-				'aria-owns': 'some-id',
-				'aria-active': 'some-active-id',
-				'not-aria-prop': 'old value',
-			}, {
-				tagName: 'div',
-				className: 'class1 class2',
-				style: {
-					backgroundColor: 'red',
-					color: 'black',
-					fontSize: '12px',
-				},
-				'aria-active': 'some-other-active-id',
-				'not-aria-prop': 'new value',
-				'aria-label': 'my label',
-			} ) ).toEqual( {
-				removedKeys: [ 'aria-owns' ],
-				updatedKeys: [ 'aria-active' ],
-			} );
-		} );
-	} );
-} );
